+++ logust/__init__.py
@@ -78,7 +78,7 @@ if TYPE_CHECKING:
     def critical(message: str, *, exception: str | None = None, **kwargs: Any) -> None: ...
 
 
-__version__ = "0.2.0"
+__version__ = "0.2.1"
 
 logger = Logger(_rust_logger)